backend/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==3.0.2
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
PyMySQL==1.1.1
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.3.2
scipy==1.15.2
setuptools==78.1.0
six==1.17.0
SQLAlchemy==2.0.39
threadpoolctl==3.6.0
typing_extensions==4.13.0
tzdata==2025.2
Werkzeug==3.1.3
wheel==0.45.1
Flask-Migrate
